# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rocky Mountain Race Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # CLUBS
  /clubs:
    get:
      operationId: listClubs
      tags: [clubs]
      summary: List all clubs
      responses:
        "200":
          description: List of clubs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Club"
    post:
      operationId: createClub
      tags: [clubs]
      summary: Create a club
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClubInput"
      responses:
        "201":
          description: Club created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
  /clubs/{clubId}:
    get:
      operationId: getClub
      tags: [clubs]
      summary: Get a club
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Club details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
    patch:
      operationId: updateClub
      tags: [clubs]
      summary: Update a club
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClubUpdate"
      responses:
        "200":
          description: Club updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"

  # EVENTS
...
  /events/{eventId}/checkins:
    get:
      operationId: listCheckins
      tags: [checkins]
      summary: List check-in status for all registered riders
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Checkin list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Checkin"
    post:
      operationId: checkinRider
      tags: [checkins]
      summary: Check in a rider
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckinInput"
      responses:
        "200":
          description: Rider checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkin"

  # RFID
...
  /dashboard/club/{clubId}:
    get:
      operationId: getClubDashboard
      tags: [dashboard]
      summary: Club dashboard summary — upcoming events, recent check-ins, total riders
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubDashboard"
  /events/{eventId}/raceday-summary:
    get:
      operationId: getRaceDaySummary
      tags: [dashboard]
      summary: Race-day live summary — checked-in count, RFID linked, motos status
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Race day summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaceDaySummary"
...
  /public/recent-results:
    get:
      operationId: listRecentResults
      tags: [dashboard]
      summary: Recent published race results across all events (public feed)
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recent results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecentResultItem"
  /reports/event/{eventId}:
    get:
      operationId: getEventReport
      tags: [reports]
      summary: Generate event report (registration, checkin, results summary)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventReport"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, club_organizer, staff, public]
        clubId:
          type: ["integer", "null"]
        createdAt:
          type: string
    Club:
      type: object
      required: [id, name, state]
      properties:
        id:
          type: integer
        name:
          type: string
        state:
          type: string
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    ClubInput:
      type: object
      required: [name, state]
      properties:
        name:
          type: string
        state:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        logoUrl:
          type: string
        website:
          type: string
        description:
          type: string
    ClubUpdate:
      type: object
      properties:
        name:
          type: string
        state:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        logoUrl:
          type: string
        website:
          type: string
        description:
          type: string
    Event:
      type: object
      required: [id, clubId, name, date, state, status]
      properties:
        id:
          type: integer
        clubId:
          type: integer
        name:
          type: string
        date:
          type: string
        state:
          type: string
        location:
          type: ["string", "null"]
        trackName:
          type: ["string", "null"]
        raceClasses:
          type: array
          items:
            type: string
        registrationOpen:
          type: ["string", "null"]
        registrationClose:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, registration_open, registration_closed, race_day, results_published, completed]
        paymentEnabled:
          type: boolean
        entryFee:
          type: ["number", "null"]
        maxRiders:
          type: ["integer", "null"]
        clubName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Checkin:
      type: object
      required: [id, eventId, riderId, riderName, checkedIn]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        riderId:
          type: integer
        riderName:
          type: string
        raceClass:
          type: string
        bibNumber:
          type: ["string", "null"]
        checkedIn:
          type: boolean
        checkedInAt:
          type: ["string", "null"]
        rfidNumber:
          type: ["string", "null"]
        rfidLinked:
          type: boolean
    CheckinInput:
      type: object
      required: [riderId]
      properties:
        riderId:
          type: integer
        rfidNumber:
          type: string
        bibNumber:
          type: string
...
    ClubDashboard:
      type: object
      required: [totalEvents, upcomingEvents, totalRiders, totalRegistrations, recentActivity]
      properties:
        totalEvents:
          type: integer
        upcomingEvents:
          type: integer
        totalRiders:
          type: integer
        totalRegistrations:
          type: integer
        checkedInToday:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
        upcomingEventList:
          type: array
          items:
            $ref: "#/components/schemas/Event"
    ActivityItem:
      type: object
      required: [type, description, timestamp]
      properties:
        type:
          type: string
        description:
          type: string
        timestamp:
          type: string
    RaceDaySummary:
      type: object
      required: [eventId, eventName, totalRegistered, checkedIn, notCheckedIn, rfidLinked, motosScheduled, motosCompleted]
      properties:
        eventId:
          type: integer
        eventName:
          type: string
        totalRegistered:
          type: integer
        checkedIn:
          type: integer
        notCheckedIn:
          type: integer
        rfidLinked:
          type: integer
        motosScheduled:
          type: integer
        motosCompleted:
          type: integer
        classSummary:
          type: array
          items:
            $ref: "#/components/schemas/ClassSummary"
    ClassSummary:
      type: object
      required: [className, registered, checkedIn]
      properties:
        className:
          type: string
        registered:
          type: integer
        checkedIn:
          type: integer
...
    RecentResultItem:
      type: object
      required: [eventId, eventName, state, date, topRider, raceClass]
      properties:
        eventId:
          type: integer
        eventName:
          type: string
        state:
          type: string
        date:
          type: string
        topRider:
          type: string
        raceClass:
          type: string
        clubName:
          type: string
    EventReport:
      type: object
      required: [eventId, eventName, type, generatedAt, data]
      properties:
        eventId:
          type: integer
        eventName:
          type: string
        type:
          type: string
        generatedAt:
          type: string
        data:
          type: object